Struct stemmer::Stemmer

pub struct Stemmer {
    // some fields omitted
}

Stemmer

Wraps Snowball's C stemming library so it can be used safely in Rust.

Examples

use stemmer::Stemmer;
let stemmer = Stemmer::new("french").unwrap();
assert_eq!("éternu", &stemmer.stem("éternuerai"));

Methods

impl Stemmer

fn list() -> Vec<&'static str>

Lists all existing algorithms, returning a vector of valid algorithms that can be used as argument for Stemmer::new

fn new(algorithm: &str) -> Option<Stemmer>

Creates a new stemmer, provided algorithm is a valid one.

Arguments

algorithm: the name of a stemming algorithm. A list of supported algorithms can be obtained with Stemmer::list()

Returns

Some(Stemmer) if algorithm exists;

None otherwise.

Examples

use stemmer::Stemmer;
let stemmer = Stemmer::new("english");
assert!(stemmer.is_some());

use stemmer::Stemmer;
let stemmer = Stemmer::new("foobar");
assert!(stemmer.is_none());

fn stem(&self, word: &str) -> String

Stems the word.

Returns an owned string.

fn stem_unsafe(&self, word: &str) -> *const str

Stems the word, unsafely.

The str pointer it returns is only valid as long as you don't call stem or stem_unsafe` again.

Unless you know what you are doing, you should use the safe stem method.

(Note that this method is actually safe to call, but derefencing the pointer it returns is unsafe)